refactor: 기본적인 HTTP 응답 설정 로직을 HttpRequestProcessor로 모음
- 모든 응답에 필요한 필드 설정 로직이 기존 HttpProcessor와 HttpRequestProcessor에 분산되어있어 HttpRequestProcessor로 모음
