+++ src/main/java/codesquad/http/HttpRequestProcessor.java
@@ -1,5 +1,8 @@
 package codesquad.http;
 
+import static codesquad.http.header.HeaderField.DATE;
+
+import codesquad.config.GlobalConstants;
 import codesquad.http.handler.DynamicRequestHandler;
 import codesquad.http.handler.DynamicRequestHandlerResolver;
 import codesquad.http.handler.RequestHandler;
@@ -7,6 +10,10 @@ import codesquad.http.handler.StaticResourceRequestHandler;
 import codesquad.http.message.HttpRequest;
 import codesquad.http.message.HttpResponse;
 import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.Locale;
+import java.util.TimeZone;
 
 public class HttpRequestProcessor {
 
@@ -25,6 +32,8 @@ public class HttpRequestProcessor {
 
         RequestHandler requestHandler = resolveRequestHandler(httpRequest.getUri().getPath());
         requestHandler.handle(httpRequest, httpResponse);
+
+        setDateHeader(httpResponse);
     }
 
     private RequestHandler resolveRequestHandler(String path) {
@@ -34,5 +43,16 @@ public class HttpRequestProcessor {
         }
 
         return staticResourceRequestHandler;
+    }
+
+    private void setDateHeader(HttpResponse httpResponse) {
+        Locale locale = GlobalConstants.getInstance().getLocale();
+        String timezone = GlobalConstants.getInstance().getTimezone();
+
+        SimpleDateFormat dateFormat = new SimpleDateFormat("EEE, dd MMM yyyy HH:mm:ss z", locale);
+        dateFormat.setTimeZone(TimeZone.getTimeZone(timezone));
+        String date = dateFormat.format(new Date());
+
+        httpResponse.setHeader(DATE.getFieldName(), date);
     }
 }
